Edited fig 5's name and its content to not show mean but show enriched rothman and cc.
